requirements.txt
streamlit==1.26.0
langchain_community==0.0.12
langchain==0.0.304
langchain_google_genai==0.0.4
PyPDF2==3.0.1
python-dotenv
